+++ pyccel/ast/builtin_methods/set_methods.py
@@ -12,9 +12,41 @@ This module contains objects which describe these methods within Pyccel's AST.
 from pyccel.ast.datatypes import NativeVoid, NativeGeneric
 from pyccel.ast.internals import PyccelInternalFunction
 
-__all__ = ('SetAdd',)
+__all__ = ('SetAdd', 'SetClear', 'SetMethod', 'SetCopy')
 
-class SetAdd(PyccelInternalFunction) :
+
+class SetMethod(PyccelInternalFunction):
+    """
+    Abstract class for set method calls.
+
+    A subclass of this base class represents calls to a specific 
+    set method.
+
+    Parameters
+    ----------
+    set_variable : TypedAstNode
+        The set on which the method will operate.
+
+    *args : TypedAstNode
+        The arguments passed to the function call.
+    """
+    __slots__ = ('_set_variable',)
+    _attribute_nodes = ('_set_variable',)
+    def __init__(self,  set_variable, *args):
+        self._set_variable = set_variable
+        super().__init__(*args)
+
+    @property
+    def set_variable(self):
+        """
+        Get the variable representing the set.
+
+        Get the variable representing the set.
+        """
+        return self._set_variable
+
+
+class SetAdd(SetMethod) :
     """
     Represents a call to the .add() method.
     
@@ -25,13 +57,12 @@ class SetAdd(PyccelInternalFunction) :
     Parameters
     ----------
     set_variable : TypedAstNode
-        The name of the set.
+        The set on which the method will operate.
 
     new_elem : TypedAstNode
         The element that needs to be added to a set.
     """
-    __slots__ = ("_set_variable", "_add_arg")
-    _attribute_nodes = ("_set_variable", "_add_arg")
+    __slots__ = ()
     _dtype = NativeVoid()
     _shape = None
     _order = None
@@ -50,24 +81,54 @@ class SetAdd(PyccelInternalFunction) :
         )
         if not is_homogeneous:
             raise TypeError("Expecting an argument of the same type as the elements of the set")
-        self._set_variable = set_variable
-        self._add_arg = new_elem
-        super().__init__()
+        super().__init__(set_variable, new_elem)
 
-    @property
-    def set_variable(self):
-        """
-        Get the variable representing the set.
 
-        Get the variable representing the set.
-        """
-        return self._set_variable
+class SetClear(SetMethod):
+    """
+    Represents a call to the .clear() method.
+    
+    The method clear is used to remove all data from a set. 
+    This operation clears all elements from the set, leaving it empty.
 
-    @property
-    def add_argument(self):
-        """
-        Get the argument which is passed to add().
+    Parameters
+    ----------
+    set_variable : TypedAstNode
+        The set on which the method will operate.
+    """
+    __slots__ = ()
+    _dtype = NativeVoid()
+    _shape = None
+    _order = None
+    _rank = 0
+    _precision = None
+    _class_type = NativeVoid()
+    name = 'clear'
 
-        Get the argument which is passed to add().
-        """
-        return self._add_arg
+    def __init__(self, set_variable):
+        super().__init__(set_variable)
+
+
+class SetCopy(SetMethod):
+    """
+    Represents a call to the .copy() method.
+
+    The copy() method in set class creates a shallow 
+    copy of a set object and returns it. 
+
+    Parameters
+    ----------
+    set_variable : TypedAstNode
+        The set on which the method will operate.
+    """
+    __slots__ = ("_dtype","_shape", "_order", "_rank", "_precision", "_class_type",)
+    name = 'copy'
+
+    def __init__(self, set_variable):
+        self._dtype = set_variable._dtype
+        self._shape = set_variable._shape
+        self._order = set_variable._order
+        self._rank = set_variable._rank
+        self._precision = set_variable._precision
+        self._class_type = set_variable._class_type
+        super().__init__(set_variable)